requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
python-multipart
openai~=1.64.0
python-dotenv~=1.0.1
pydantic~=2.10.6
jinja2
DateTime~=5.5
pyttsx3~=2.98
requests~=2.32.3
SpeechRecognition~=3.14.1
beautifulsoup4~=4.13.3